hello_agent/async-hello-agent/pyproject.toml
[project]
name = "async-hello-agent"
version = "0.1.0"
description = "An async hello agent using OpenAI Agents SDK"
authors = [
    {name = "Your Name"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "openai-agents (>=0.0.16,<0.0.17)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "openai (>=1.82.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
